7.3.8 Refactor tests to be maintainable
* Moved all tests into their own module(s) * Cleaned up all unused
imports * This makes the entire test suite build as a single binary
which can speed up testing, since multiple files that are not in a
module are built in parallel, but are linked sequentially. This can save
up to 1.9x time! * Also this lets us encapsulate our test helper logic
into the helpers function. * We can also refactor each file into a
module as complexity grows, we could even make helper functions for just
one module like subscriptions as it grows.