7.7.2.1 Red test - Make sure we are sending an email link
* Added linkify to search for links within a string. * used the
recieved_requests method on the mock server to get all the requests sent
to it. * parse the body as json from the slice of bytes of the body into
a serde_json::value type which we can convert to a string and then pass
to linkify to search for links. * we use an anonymous function/closure
in the test to reduce repetition in our tests.